Reorganize ETHProvider
