+++ electrum_gui/common/provider/chains/cfx/conflux.py
@@ -11,7 +11,7 @@ from electrum_gui.common.basic.request.exceptions import JsonRPCException
 from electrum_gui.common.basic.request.json_rpc import JsonRPCRequest
 from electrum_gui.common.provider.chains.cfx.sdk import cfx_address, consts
 from electrum_gui.common.provider.chains.cfx.sdk.types import Drip, SponsorInfo
-from electrum_gui.common.provider.chains.eth.clients import utils
+from electrum_gui.common.provider.chains.eth.clients import helper
 from electrum_gui.common.provider.chains.eth.clients.geth import (
     InvalidContractAddress,
     _extract_eth_call_str_result,
@@ -164,7 +164,7 @@ class CFXClient(ClientInterface, BatchGetAddressMixin):
             json_response = e.json_response
             if isinstance(json_response, dict) and "error" in json_response:
                 error_message = json_response.get("error", {}).get("message") or ""
-                utils.handle_broadcast_error(error_message)
+                helper.raise_broadcast_error(error_message)
 
             raise e
